% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-file-helpers.R
\name{clear_test_logs}
\alias{clear_test_logs}
\title{Clears out all test logs from given folder}
\usage{
clear_test_logs(folder)
}
\arguments{
\item{folder}{folder which to clean test logs from}
}
\description{
Clears out all test logs from given folder
}
