% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-preprocessing.R
\name{convert_action}
\alias{convert_action}
\title{Converts actions written in Czech to english counterparts}
\usage{
convert_action(exp)
}
\arguments{
\item{exp}{experiment_log}
}
\value{
modified experiment log
}
\description{
Converts actions written in Czech to english counterparts
}
