% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-preprocessing.R
\name{convert_name_to_item_code}
\alias{convert_name_to_item_code}
\title{Converts existing}
\usage{
convert_name_to_item_code(items, language)
}
\arguments{
\item{language}{what language to use for conversions. e.g. "CZR".
This language needs to figure out in the intrm_translations.rda file}
}
\value{
vector of converted values
The [1] is there because some of the items have two IDs
in the item_translations data
}
\description{
Converts existing
}
