% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-preprocessing.R
\name{convert_rightwrong}
\alias{convert_rightwrong}
\title{Converts RightWrong written in Czech to english counterparts.
Then it converts the column to logical}
\usage{
convert_rightwrong(exp)
}
\arguments{
\item{exp}{experiment_log}
}
\value{
modified experiment log
}
\description{
Converts RightWrong written in Czech to english counterparts.
Then it converts the column to logical
}
