% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-getters.R
\name{get_finished_trials_indices.supermarket}
\alias{get_finished_trials_indices.supermarket}
\title{Returns indices}
\usage{
get_finished_trials_indices.supermarket(obj)
}
\arguments{
\item{obj}{supermarket object}
}
\value{
indices e.g. TestCycles of trials which were finished
}
\description{
Returns indices
}
