% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-getters.R
\name{get_pickup_position}
\alias{get_pickup_position}
\title{Returns position data for picking up a specific object in given trial}
\usage{
get_pickup_position(obj, object_name = NULL, i_cycle = NULL)
}
\arguments{
\item{obj}{loaded supermarket data}

\item{object_name}{object name as it appears in the experiment log}

\item{i_task}{if there are more objects of the same name picked up,
this defines which cycle you are interested in}
}
\value{
navr object with the object path (from previous item to the
current item)
}
\description{
Returns position data for picking up a specific object in given trial
}
