% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-getters.R
\name{get_trial_position.supermarket}
\alias{get_trial_position.supermarket}
\title{Returns position during trial}
\usage{
\method{get_trial_position}{supermarket}(obj, i_trial)
}
\arguments{
\item{obj}{supermarket object}

\item{i_trial}{}
}
\value{
navr position object
}
\description{
Returns position during trial
}
