% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-getters.R
\name{get_trial_times.supermarket}
\alias{get_trial_times.supermarket}
\title{Returns trial times}
\usage{
\method{get_trial_times}{supermarket}(obj, i_trial)
}
\arguments{
\item{i_trial}{}
}
\description{
Returns trial times
}
