% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-preprocessing.R
\name{has_item_codes}
\alias{has_item_codes}
\title{Checks if the items in the data are coded with codes or with czech names}
\usage{
has_item_codes(obj)
}
\arguments{
\item{obj}{}
}
\value{
logical value
}
\description{
Checks if the items in the data are coded with codes or with czech names
}
