% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-categories.R
\docType{data}
\name{item_categories}
\alias{item_categories}
\title{Item categories encoding table}
\format{A data.frame with 212 items and 5 variables:
\describe {
   \item{ID}{ID of the item. In english, but ultimately convertable}
   \item{CZ}{Czech name of the item. UTF-8 encoded}
   \item{EN}{English common name of the item. uTF-8 encoded}
   \item{CZR}{}
   \item{Category}{name of the category ofg the item in english}
}}
\usage{
item_categories
}
\description{
This dataframe serves the purpose to convert between language versions
as well as codument categorization of individual objects
}
\keyword{datasets}
