% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-item-translations.R
\docType{data}
\name{item_translations}
\alias{item_translations}
\title{Item translation}
\format{A data.frame with 212 items and 5 variables:
\describe {
   \item{ID}{ID of the item}
   \item{CZ}{Czech name of the item. UTF-8 encoded}
   \item{EN}{English common name of the item. uTF-8 encoded}
   \item{CZR}{}
}}
\usage{
item_translations
}
\description{
This dataframe serves the purpose to convert between language versions
}
\details{
Language options at this point include "CZ", "CZR" and "EN"
}
\keyword{datasets}
