% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-loading.R
\name{load_supermarket_settings}
\alias{load_supermarket_settings}
\title{Loads .json settings file}
\usage{
load_supermarket_settings(filepath)
}
\arguments{
\item{filepath}{path to the .json settings file. In newer versions, settings is usually already included in the header,
but in older versions it needs to be loaded sparately.}
}
\value{
list with loaded settings
}
\description{
Loads .json settings file
}
