% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-visualisations.R
\name{plot_supermarket_path}
\alias{plot_supermarket_path}
\title{Plots position data and adds background}
\usage{
plot_supermarket_path(obj, background = FALSE,
  background_type = "categories", custom_background = NULL, ...)
}
\arguments{
\item{obj}{navr object filtered from the supermarket data position}

\item{background}{do you want to add background to the path}

\item{background_type}{if you want installed background (categories, plain)}

\item{custom_background}{path to your custom background}
}
\value{
ggplot object
}
\description{
Plots position data and adds background
}
