% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-file-helpers.R
\name{rename_action_to_test_logs}
\alias{rename_action_to_test_logs}
\title{Convets action logs to test logs. The outcome is the version 5
of the supermarket data due these files originating in the pre version 5 era}
\usage{
rename_action_to_test_logs(folder)
}
\arguments{
\item{folder}{folder to edit}
}
\description{
Convets action logs to test logs. The outcome is the version 5
of the supermarket data due these files originating in the pre version 5 era
}
