% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-analysis-items.R
\name{task_performance_all}
\alias{task_performance_all}
\title{Calculates results for each tríal}
\usage{
task_performance_all(obj)
}
\arguments{
\item{obj}{supermarket experiment loaded with load_supermarket_experiment. Needs to have tasklist attached}
}
\description{
Calculates results for each tríal
}
