% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-analysis-items.R
\name{task_performance_trial}
\alias{task_performance_trial}
\title{Calculates results for each tríal}
\usage{
task_performance_trial(obj, i_trial)
}
\arguments{
\item{obj}{supermarket experiment loaded with load_supermarket_experiment. Needs to have tasklist attached}

\item{i_trial}{which trial number as it stands in the tasklist and "TestCycle"}
}
\description{
Calculates results for each tríal
}
