% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-getters.R
\name{get_trial_wanted_items}
\alias{get_trial_wanted_items}
\title{Returns list of correct items from a trial}
\usage{
get_trial_wanted_items(obj, i_trials)
}
\arguments{
\item{i_trial}{}
}
\description{
this function became necessary when the entire structure of the task changed and
suddenly the list of items is no longer serialised and is generated but never noted what those
items are.

This function requires both actions and results log to be present because it needs both to
"deduce" which items were supposed to be collected. In case the tasklist is present, the items
are collected from the tasklist instead
}
