% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-loading.R
\name{load_supermarket_experiments}
\alias{load_supermarket_experiments}
\title{Loads supermarket experiment data from folder}
\usage{
load_supermarket_experiments(folder)
}
\arguments{
\item{folder}{folder with one or more supermarket experiments}
}
\value{
list with loaded experiments
}
\description{
Loads supermarket experiment data from folder
}
