% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-loading.R
\name{load_supermarket_experiments}
\alias{load_supermarket_experiments}
\title{Loads supermarket experiment data from folder}
\usage{
load_supermarket_experiments(folder, language = "CZ", ...)
}
\arguments{
\item{folder}{folder with one or more supermarket experiments}

\item{language}{language of the items in the experiment log.
Only important if you are not logging item codes
See language options in [item_translations]. Default is "CZ".}

\item{...}{other parameters being set to brainvr.reader::load_experiments.
typically this is `override`, and `save`}
}
\value{
list with loaded experiments
}
\description{
Loads supermarket experiment data from folder
}
